backend/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# 缓存和任务队列
redis==5.0.1
celery==5.3.4

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 文件处理
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0

# AI 服务
openai==1.7.2
anthropic==0.40.0

# 工具
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
